+++ src/backend_skia/mod.rs
@@ -64,8 +64,8 @@ impl Render for Backend {
 }
 
 impl OutputImage for skia::Surface {
-    fn save(
-        &self,
+    fn save_png(
+        &mut self,
         path: &std::path::Path,
     ) -> bool {
         self.save(path.to_str().unwrap())